Toll Like Receptor Model with MyD88 Activation
.ipynb file should be run within this folder, with reactions.csv , rateLaws.csv , parameters.csv , and csv2model.py
The ipynb file will assemble the model definition files into a new model file such as justTLR_MyD88_SELFACTIVATING.jl , then run it as a simulation.
The programming language is Julia.
The .ipynb file format is a Jupyter Notebook. (This comes with the default anaconda Navigator installation)
For quick reference, justTLR_MyD88_SELFACTIVATING_example.jl is attached, which is a pre-assembled model file.